[REFACTOR]:Implements interfaces package
