+++ packages/afp/lib/afp.dart
@@ -1,5 +1,4 @@
 /// A Very Good Project created by Very Good CLI.
 library;
 
-export 'src/base_error.dart';
-export 'src/result.dart'
+export 'src/afp.dart';
