+++ packages/afp/lib/src/interfaces/base_error.dart
